# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My docs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/transfer/recruitment/update": {
            "post": {
                "tags": [
                    "Projects Transfer"
                ],
                "summary": "Обновить состояние проектов на 'Набор'",
                "description": "Этот метод обновляет состояние проектов, которые находятся в состоянии 'Одобрено' (state_id = 9), на состояние 'Набор' (state_id = 1). Он также обновляет состояние кандидатов в соответствующих проектах.",
                "operationId": "updateToRecruitmentState",
                "responses": {
                    "200": {
                        "description": "Успешно обновлены проекты",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updated_projects": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID проекта",
                                                        "type": "integer"
                                                    },
                                                    "state_id": {
                                                        "description": "ID состояния проекта",
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "updated_count": {
                                            "description": "Количество обновленных проектов",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при загрузке или обработке данных"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
...
        "/api/supervisors/{id}": {
            "get": {
                "tags": [
                    "Supervisor"
                ],
                "summary": "Получение информации о преподавателе по ID",
                "description": "Получение информации о преподавателе по ID",
                "operationId": "266d09407e0a5c86ba417cb9b493f75e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID преподавателя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о преподавателе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supervisor"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Department": {
                "properties": {
                    "id": {
                        "description": "ID Кафедры",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название кафедры",
                        "type": "string"
                    },
                    "institute": {
                        "$ref": "#/components/schemas/Institute"
                    }
                },
                "type": "object"
            },
...
            "Institute": {
                "properties": {
                    "id": {
                        "description": "ID института",
                        "type": "integer"
                    },
                    "maxApprovedProjects": {
                        "description": "Максимальное кол-во одобренных проектов",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название института",
                        "type": "string"
                    },
                    "maxSpringApprovedProjects": {
                        "description": "Максимальное кол-во одобренных проектов на весну",
                        "type": "integer"
                    },
                    "maxAutumnApprovedProjects": {
                        "description": "Максимальное кол-во одобренных проектов на осень",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "Supervisor": {
                "properties": {
                    "id": {
                        "description": "ID Преподавателя",
                        "type": "integer"
                    },
                    "fio": {
                        "description": "ФИО",
                        "type": "string"
                    },
                    "email": {
                        "description": "Почта",
                        "type": "string"
                    },
                    "about": {
                        "description": "Информация о преподавателе",
                        "type": "string"
                    },
                    "position": {
                        "description": "Должность",
                        "type": "string"
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    },
                    "created_at": {
                        "description": "Дата создания записи в БД",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "Дата обновления записи в БД",
                        "type": "string"
                    },
                    "canReviewProjects": {
                        "description": "Может ли делать ревью проектов",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}